requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-mptt==0.12.0
idna==2.10
oauthlib==3.1.0
Pillow==8.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.3
